+++ src/App/Fossa/Analyze/Debug.hs
@@ -44,11 +44,13 @@ import Control.Effect.Stack (Stack (Context))
 import Control.Effect.Sum (Member, inj)
 import Control.Monad.IO.Class (MonadIO)
 import Data.Aeson (ToJSON (toEncoding), defaultOptions, genericToEncoding)
+import Data.Aeson qualified as Aeson
 import Data.Bifunctor (bimap)
 import Data.Map qualified as Map
 import Data.String.Conversion (toText)
 import Data.Text (Text, toLower)
 import Data.Word (Word64)
+import Diag.Result (Result (..))
 import Effect.Exec (Exec, ExecF (..))
 import Effect.Logger (Logger, LoggerF (..))
 import Effect.ReadFS (ReadFS, ReadFSF (..))
@@ -107,23 +109,30 @@ collectDebugBundle ::
   , Has Diagnostics sig m
   , Has Logger sig m
   , Has (Lift IO) sig m
+  , ToJSON a
   ) =>
   cfg ->
-  DebugEverythingC (RecordC ReadFSF (RecordC ExecF (DebugC m))) a ->
-  m (DebugBundle cfg, a)
+  DebugEverythingC (RecordC ReadFSF (RecordC ExecF (DebugC m))) (Result a) ->
+  m (DebugBundle cfg, (Result a))
 collectDebugBundle cfg act = do
   (scope, (execJournal, (readFSJournal, res))) <- runDebug . runRecord @ExecF . runRecord @ReadFSF . debugEverything $ act
   sysInfo <- sendIO collectSystemInfo
   args <- sendIO getCommandArgs
   envVars <- collectEnvVariables
+
+  let output :: Aeson.Value = case res of
+        Failure _ _ -> "scan was not successful, no output"
+        Success _ a -> Aeson.toJSON a
+
   let bundle =
         DebugBundle
-          { bundleScope = scope
-          , bundleSystem = sysInfo
+          { bundleSystem = sysInfo
           , bundleCLIVersion = fullVersionDescription
           , bundleArgs = args
           , bundleConfig = cfg
           , bundleEnvVariables = envVars
+          , bundleOutput = output
+          , bundleScope = scope
           , bundleJournals =
               BundleJournals
                 { bundleJournalReadFS = readFSJournal
@@ -152,12 +161,13 @@ collectSystemInfo = do
       systemMemory
 
 data DebugBundle cfg = DebugBundle
-  { bundleScope :: Scope
-  , bundleSystem :: SystemInfo
+  { bundleSystem :: SystemInfo
   , bundleCLIVersion :: Text
   , bundleArgs :: [Text]
   , bundleConfig :: cfg
   , bundleEnvVariables :: Map.Map Text Text
+  , bundleOutput :: Aeson.Value
+  , bundleScope :: Scope
   , bundleJournals :: BundleJournals
   }
   deriving (Show, Generic)