+++ vendor_download.sh
@@ -79,7 +79,7 @@ rm $WIGGINS_RELEASE_JSON
 echo "Wiggins download successful"
 echo
 
-THEMIS_TAG="2023-04-17-24f42c2-1681765273"
+THEMIS_TAG="2023-04-25-95b18b6-1682456045"
 echo "Downloading themis binary"
 echo "Using themis release: $THEMIS_TAG"
 THEMIS_RELEASE_JSON=vendor-bins/themis-release.json
@@ -117,9 +117,8 @@ echo
 echo "Marking binaries executable"
 chmod +x vendor-bins/*
 
-echo "Compressing binaries"
+echo "Compressing index.gob"
 xz vendor-bins/index.gob
-find vendor-bins -type f -not -name '*.xz' | xargs upx || echo "WARN: 'upx' command not found, binaries will not be compressed"
 
 echo "Vendored binaries are ready for use"
 ls -lh vendor-bins/